Share env urls
