+++ vscode/src/configuration.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from 'vitest'
 import type * as vscode from 'vscode'
 
-import { DOTCOM_URL } from './chat/protocol'
+import { DOTCOM_URL } from '@sourcegraph/cody-shared/src/sourcegraph-api/environments'
+
 import { getConfiguration } from './configuration'
 
 describe('getConfiguration', () => {
